+++ octave/ConcentrationDescriptor11_12.m
@@ -1,6 +1,6 @@
 function [bc50] = ConcentrationDescriptor11_12(logK11, logK12)
   %this function calculates all concentrational descriptors for any given logK21 and logK11 pairs
-  
+  printf("\n\n\n");
   global b11;
   global b12;
   
@@ -21,31 +21,32 @@ function [bc50] = ConcentrationDescriptor11_12(logK11, logK12)
   printf("B in Solution: %d\n", B*10^6);
   printf("AB in Solution: %d\n", AB*10^6);
   printf("AB2 in Solution: %d\n", AB2*10^6);
- B0 = B + AB + 2 *AB2;
-   printf("Lets calculate the relative concentrations from the approximated soluation\n");
+  B0 = B + AB + 2 *AB2;
+  printf("Lets calculate the relative concentrations from the approximated soluation\n");
   printf("B in Solution: %d\n", B/B0);
   printf("AB in Solution: %d\n", AB/B0);
   printf("A2B in Solution: %d\n", 2*AB2/B0);
   printf("Lets do it the inverse way \n");
   
-  %B = quad("BFunction11_12", lower_limit, upper_limit);
+  B = quad("BFunction11_12", lower_limit, upper_limit);
   AB = quad("ABFunction11_12", lower_limit, upper_limit);
   AB2 = quad("AB2Function11_12", lower_limit, upper_limit);
-  AAB2 = quad("AAB2Function11_12", lower_limit, upper_limit)
-
-  B = AAB2;
-  (AAB2-AB)/2
-  printf("The normal way round ... \n");
-  %printf("A in Solution: %d\n", A*10^6);
+  AAB2 = quad("AAB2Function11_12", lower_limit, upper_limit);
+  %AB + 2*AB2 - AAB2
+  %B = AAB2;
+  %(AAB2-AB)/2
+  %printf("The normal way round ... \n");
+  printf("A in Solution: %d\n", A*10^6);
   printf("B in Solution: %d\n", B*10^6);
   printf("AB in Solution: %d\n", AB*10^6);
   printf("AB2 in Solution: %d\n", AB2*10^6);
   
-  B0 = B + AB + 2 *AB2;
-   printf("Lets calculate the relative concentrations from the approximated soluation\n");
+  B0 = B + AB + 2 *AB2
+  A0 = A + AB + AB2
+  printf("Lets calculate the relative concentrations from the approximated soluation\n");
   printf("B in Solution: %d\n", B/B0);
   printf("AB in Solution: %d\n", AB/B0);
-  printf("A2B in Solution: %d\n", 2*AB2/B0);
+  printf("AB2 in Solution: %d\n", 2*AB2/B0);
   printf("Lets do it the inverse way \n");
   
 endfunction