+++ src/core/minimizer.cpp
@@ -101,7 +101,7 @@ int NonLinearFitThread::NonLinearFit()
 void NonLinearFitThread::Print(const QString& message)
 {
 #ifdef _DEBUG
-    qDebug() << message;
+//    qDebug() << message;
 #else
     Q_UNUSED(message)
 #endif